+++ src/ros_mqtt_bridge/task_type.py
@@ -1,4 +1,10 @@
 from dataclasses import dataclass, field
+from enum import Enum, auto
+
+
+class CommunicationType(Enum):
+    TOPIC2TOPIC = auto()
+    TOPIC2SERVICE = auto()
 
 
 @dataclass
@@ -15,3 +21,6 @@ class ROS2MQTTItem:
     last_received_time: float = field(default=0.0)
     msg: object = field(default=None)
     converter: str = field(default="")
+    commmunication_type: CommunicationType = field(
+        default=CommunicationType.TOPIC2TOPIC
+    )