+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package ros_mqtt_bridge
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.2.0 (2024-01-07)
+------------------
+* Fix bug for multiple ROS Service response
+* Fix bug if config value is empty for topic2topic and topic2service
+* Add timeout for service call to terminate thread if service call takes too long
+
 0.1.0 (2023-12-20)
 ------------------
 * Initial implementation of `ros_mqtt_bridge` package